+++ src/api/projects/access_requests.rs
@@ -1,0 +1,31 @@
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+//! Project access requests API endpoints.
+//!
+//! These endpoints are used for querying projects access requests
+
+mod access_requests;
+mod approve;
+mod deny;
+mod request;
+
+pub use self::access_requests::ProjectAccessRequests;
+pub use self::access_requests::ProjectAccessRequestsBuilder;
+pub use self::access_requests::ProjectAccessRequestsBuilderError;
+
+pub use self::approve::ProjectAccessLevel;
+pub use self::approve::ProjectAccessRequestsApprove;
+pub use self::approve::ProjectAccessRequestsApproveBuilder;
+pub use self::approve::ProjectAccessRequestsApproveBuilderError;
+
+pub use self::deny::ProjectAccessRequestsDeny;
+pub use self::deny::ProjectAccessRequestsDenyBuilder;
+pub use self::deny::ProjectAccessRequestsDenyBuilderError;
+
+pub use self::request::ProjectAccessRequest;
+pub use self::request::ProjectAccessRequestBuilder;
+pub use self::request::ProjectAccessRequestBuilderError;